refactor(run0): address inline code review comments
Code improvements:
- Remove redundant ensure_non_empty calls (jaccard_similarity handles this)
- Fix column computation to use UTF-16 code units for SARIF compatibility
- Split tests.rs into tests.rs and tests_emit.rs (now under 400 lines)
- Replace unwrap_or_else with expect in pair() test helper
- Import SimilarityRatio directly in types.rs (remove super:: prefix)
- Use env!("CARGO_PKG_VERSION") in run0_sarif_behaviour.rs

Documentation improvements:
- Wrap long shell commands in execplan to respect 120-column limit
- Fix article grammar: 'a SHA-256' -> 'an SHA-256' in design doc
- Clarify error types across layers (IndexError vs Run0Error)

The UTF-16 column fix ensures SARIF viewers correctly highlight code
in files with multi-byte UTF-8 characters, as SARIF columns are
UTF-16 code units per the specification.
